Rename "pages" package to "views"
