+++ src/main/java/website/fgto/royals_of_britain/views/Footer.html
