+++ src/main/java/website/fgto/royals_of_britain/views/houses/intros/Stuart.java
@@ -1,11 +1,11 @@
-package website.fgto.royals_of_britain.pages.houses.intros;
+package website.fgto.royals_of_britain.views.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Wessex extends Panel {
+public class Stuart extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Wessex(String id) {
+  public Stuart(String id) {
     super(id);
   }
 }
