+++ src/test/java/website/fgto/TestHomePage.java
@@ -3,7 +3,7 @@ package website.fgto;
 import org.apache.wicket.util.tester.WicketTester;
 import org.junit.Before;
 import org.junit.Test;
-import website.fgto.royals_of_britain.pages.HomePage;
+import website.fgto.royals_of_britain.views.HomePage;
 import website.fgto.royals_of_britain.WicketApplication;
 
 /**
